requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Flask==2.2.3
paho-mqtt==1.6.1
supabase==1.0.2
